requirements.txt
# Core agent framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-aws>=0.2.0
langchain-community>=0.3.0
langsmith>=0.1.0

# AWS SDK
boto3>=1.35.0
botocore>=1.35.0
aws-cdk-lib>=2.160.0
constructs>=10.3.0

# Data & validation
pydantic>=2.7.0
faker>=26.0.0
pandas>=2.2.0
numpy>=1.26.0

# Governance & policy
# Note: presidio-analyzer/anonymizer require Python >=3.10 (spacy dependency).
# They are optional — pii_scrubber.py falls back to regex-only when not installed.
# Install via: pip install -r requirements-optional.txt
# OPA: run `opa run` with policies/ — HTTP API evaluated via httpx in policy_engine.py
PyJWT[crypto]>=2.8.0

# Observability
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
watchtower>=3.3.0

# API & serving
fastapi>=0.115.0
uvicorn>=0.31.0
mangum>=0.19.0

# Utilities
python-dotenv>=1.0.0
tenacity>=9.0.0
structlog>=24.4.0
rich>=13.9.0
typer>=0.12.0
httpx>=0.27.0

# MCP client — LangGraph mcp_tools node when CALCLAIM_MCP_URL is set (skipped on Python <3.10)
mcp>=1.2.0,<2; python_version >= "3.10"
